+++ backend/src/main/java/com/wandr/backend/dto/traveller/UpdateProfileDTO.java
@@ -10,7 +10,7 @@ public class UpdateProfileDTO {
     private String email;
     private String country;
     private String profileImage;  // This will be the URL or filename of the profile image
-    private List<Integer> categories;
-    private List<Integer> activities;
+    private List<Long> categories;
+    private List<Long> activities;
     private String membership;
 }
